Make history more intuitive
When the user returns back to the main menu from the maze page or from
the settings page, it should feel as a 'back' action in history.
Previously, if the user opened the maze page, and then returned to the
main menu through the 'return to main menu' button, and then back
through the browser's back button, they would return to the maze, which
was irritating.